ci: don't store $CARGO_HOME in Github Actions cache (#443)
There currently is a problem related to Github Actions in general or the
'actions-rs/cargo' action in particular, which causes cargo to be unable
to find the 'serde_derive' crate when the cargo home dir has been
restored from cache.

This issue is also being reported by others, e.g. at
https://github.com/actions-rs/cargo/issues/111.