Add set_flags_from_command_line_with_usage fn
This commit suggests adding a usage string which can be passed to V8
when setting the command line flags. If the --help flag is specified
then the usage string will be printed before the V8 options.

The motivation for this would be to enable embedders of rusty_v8 to
pass their own usage string and have their usage printed before V8's
options.

For example:

Usage: binaryname --startup-src=file

SSE3=1 SSSE3=1 SSE4_1=1 SSE4_2=1 SAHF=1 AVX=1 FMA3=1 BMI1=1 BMI2=1
The following syntax for options is accepted (both '-' and '--' are ok):
  --flag        (bool flags only)
  --no-flag     (bool flags only)
  --flag=value  (non-bool flags only, no spaces around '=')
  --flag value  (non-bool flags only)
  --            (captures all remaining args in JavaScript)

Options:
...