Attempted to implement `Locker` for `Isolate`.
This adds a wrapper `SharedIsolate` for  `OwnedIsolate` (this way all the support code in OwnedIsolate doesn't need to be re-created).

Originally an Unlocker was added as well, but an Unlocker being used inside a `Locker` would result in HandleScope instance failing to be created once the Unlocker dropped.

This may be indicating that some part of the Locker api itself isn't working properly with rusty_v8's OwnedIsolate overhead, but the example code does appear to work as expected.

As `Unlocker` isn't a direct needed in my use case, it was removed in lue of implementing a bare minimum locker.