Always run `gn gen`
As described in #1831, the `build.rs` script currently skips the run of
`gn gen` if the `build.ninja` file already exists in the GN directory,
i.e. if GN has been run before. This breaks if we're passing extra
arguments to GN through the `GN_ARGS` evironment variable. These values
aren't taken into account then.

This PR updates the build script to always run `gn gen` which makes sure
that the arguments in the `gn_out` dir are updated.

Fixes: #1831